fix: gamma 全链路 Double 化，修复 OSD 显示 nan

+++ iina/OSDMessage.swift
@@ -105,7 +105,7 @@ enum OSDMessage {
   case hue(Int)
   case saturation(Int)
   case brightness(Int)
-  case gamma(Int)
+  case gamma(Double)
 
   case addFilter(String)
   case removeFilter
